move code into protocol extension
